% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native_role_groups.R
\name{add_roles_to_group}
\alias{add_roles_to_group}
\alias{remove_role_from_group}
\title{Add/Remove role to/from group}
\usage{
add_roles_to_group(group, role, dataverse, key = Sys.getenv("DATAVERSE_KEY"),
  server = Sys.getenv("DATAVERSE_SERVER"), ...)

remove_role_from_group(group, role, dataverse,
  key = Sys.getenv("DATAVERSE_KEY"),
  server = Sys.getenv("DATAVERSE_SERVER"), ...)
}
\arguments{
\item{group}{An object of class \dQuote{dataverse_group}, or a character string containing the \dQuote{alias} for a group. If a character string, \code{dataverse} must also be specified as the alias is local to a given Dataverse.}

\item{role}{A character string specifying a role, or an object of class \dQuote{dataverse_role}.}

\item{dataverse}{A character string specifying a Dataverse name or an object of class \dQuote{dataverse}. Can be missing if \code{role} is an object of class \dQuote{dataverse_group}.}

\item{key}{A character string specifying a Dataverse server API key. If one is not specified, functions calling authenticated API endpoints will fail. Keys can be specified atomically or globally using \code{Sys.setenv("DATAVERSE_KEY" = "examplekey")}.}

\item{server}{A character string specifying a Dataverse server. There are multiple Dataverse installations, but the defaults is to use the Harvard Dataverse. This can be modified atomically or globally using \code{Sys.setenv("DATAVERSE_SERVER" = "dataverse.example.com")}.}

\item{...}{Additional arguments passed to an HTTP request function, such as \code{\link[httr]{GET}}, \code{\link[httr]{POST}}, or \code{\link[httr]{DELETE}}.}
}
\value{
A list.
}
\description{
Add/remove role from group
}
\details{
Role groups can be used to organize roles for a Dataverse. These functions add or remove roles from a group
}
