% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native_user.R
\name{create_user}
\alias{create_user}
\title{Create user}
\usage{
create_user(password, key = Sys.getenv("DATAVERSE_KEY"),
  server = Sys.getenv("DATAVERSE_SERVER"), ...)
}
\arguments{
\item{password}{A character vector specifying the password for the new user.}

\item{key}{A character string specifying a Dataverse server API key. If one is not specified, functions calling authenticated API endpoints will fail. Keys can be specified atomically or globally using \code{Sys.setenv("DATAVERSE_KEY" = "examplekey")}.}

\item{server}{A character string specifying a Dataverse server. There are multiple Dataverse installations, but the defaults is to use the Harvard Dataverse. This can be modified atomically or globally using \code{Sys.setenv("DATAVERSE_SERVER" = "dataverse.example.com")}.}

\item{...}{Additional arguments passed to an HTTP request function, such as \code{\link[httr]{GET}}, \code{\link[httr]{POST}}, or \code{\link[httr]{DELETE}}.}
}
\value{
A list.
}
\description{
Create a Dataverse user
}
\details{
Create a new Dataverse user.
}
\examples{
\dontrun{
create_user("password")
}
}
\seealso{
\code{\link{get_user_key}}
}

