\name{dataverse-package}
\alias{dataverse-package}
\alias{dataverse}
\docType{package}
\title{Client for Dataverse 4 Repositories}
\description{Provides access to Dataverse 4 APIs, enabling data search, retrieval, and deposit.}
\author{
Thomas J. Leeper

Maintainer: Thomas J. Leeper <thosjleeper@gmail.com>
}
\references{
\href{http://guides.dataverse.org/en/latest/api/index.html}{Dataverse API Documentation}

\href{http://dataverse.org/}{Dataverse Homepage}

\href{https://dataverse.harvard.edu/}{Harvard IQSS Dataverse}
}
\keyword{ package }
%\seealso{}
%\examples{}
