% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataAccess.R
\name{dataverse_file}
\alias{dataverse_file}
\title{Download File(s)}
\usage{
dataverse_file(file, doi = NULL, format = c("original", "RData", "prep",
  "bundle"), vars = NULL, key = Sys.getenv("DATAVERSE_KEY"),
  server = Sys.getenv("DATAVERSE_SERVER"), ...)
}
\arguments{
\item{file}{An integer specifying a file identifier or, if \code{doi} is specified, a character string specifying a file name within the DOI-identified dataset.}

\item{doi}{If a DOI (or handle) is supplied, then \code{file} can simply be the filename rather than Dataverse file ID number.}

\item{format}{A character string specifying a file format. By default, this is \dQuote{original} (the original file format). If \dQuote{RData} or \dQuote{prep} is used, an alternative is returned. If \dQuote{bundle}, a compressed directory containing a bundle of file formats is returned.}

\item{vars}{A character vector specifying one or more variable names, used to extract a subset of the data.}

\item{key}{A character string specifying a Dataverse server API key. If one is not specified, functions calling authenticated API endpoints will fail. Keys can be specified atomically or globally using \code{Sys.setenv("DATAVERSE_KEY" = "examplekey")}.}

\item{server}{A character string specifying a Dataverse server. There are multiple Dataverse installations, but the defaults is to use the Harvard Dataverse. This can be modified atomically or globally using \code{Sys.setenv("DATAVERSE_SERVER" = "dataverse.example.com")}.}

\item{...}{Additional arguments passed to an HTTP request function, such as \code{\link[httr]{GET}}.}
}
\value{
A list.
}
\description{
Download Dataverse File(s)
}
\details{
This function provides access to data files from a Dataverse entry.
}
\examples{
\dontrun{}
}

