% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/native_dv.R
\name{delete_dataverse}
\alias{delete_dataverse}
\title{Delete Dataverse}
\usage{
delete_dataverse(dataverse, key = Sys.getenv("DATAVERSE_KEY"),
  server = Sys.getenv("DATAVERSE_SERVER"), ...)
}
\arguments{
\item{dataverse}{A character string specifying a Dataverse name or an object of class \dQuote{dataverse}.}

\item{key}{A character string specifying a Dataverse server API key. If one is not specified, functions calling authenticated API endpoints will fail. Keys can be specified atomically or globally using \code{Sys.setenv("DATAVERSE_KEY" = "examplekey")}.}

\item{server}{A character string specifying a Dataverse server. There are multiple Dataverse installations, but the defaults is to use the Harvard Dataverse. This can be modified atomically or globally using \code{Sys.setenv("DATAVERSE_SERVER" = "dataverse.example.com")}.}
}
\value{
A logical.
}
\description{
Delete a dataverse
}
\details{
This function deletes a Dataverse.
}
\examples{
\dontrun{}
}
\seealso{
\code{\link{create_dataverse}}, \code{\link{update_dataverse}}, \code{\link{get_dataverse}}, \code{\link{publish_dataverse}}, \code{\link{delete_dataset}}
}

