% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataframe.R
\name{get_dataframe_internal}
\alias{get_dataframe_internal}
\title{Write to temp and apply function}
\usage{
get_dataframe_internal(raw, filename, .f)
}
\description{
Write to temp and apply function
}
\keyword{internal}
