% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native_role_groups.R
\name{get_group}
\alias{get_group}
\title{Get role group}
\usage{
get_group(group, dataverse, key = Sys.getenv("DATAVERSE_KEY"),
  server = Sys.getenv("DATAVERSE_SERVER"), ...)
}
\arguments{
\item{group}{An object of class \dQuote{dataverse_group}, or a character string containing the \dQuote{alias} for a group. If a character string, \code{dataverse} must also be specified as the alias is local to a given Dataverse.}

\item{dataverse}{A character string specifying a Dataverse name or an object of class \dQuote{dataverse}. Can be missing if \code{role} is an object of class \dQuote{dataverse_group}.}

\item{key}{A character string specifying a Dataverse server API key. If one is not specified, functions calling authenticated API endpoints will fail. Keys can be specified atomically or globally using \code{Sys.setenv("DATAVERSE_KEY" = "examplekey")}.}

\item{server}{A character string specifying a Dataverse server. There are multiple Dataverse installations, but the defaults is to use the Harvard Dataverse. This can be modified atomically or globally using \code{Sys.setenv("DATAVERSE_SERVER" = "dataverse.example.com")}.}

\item{...}{Additional arguments passed to an HTTP request function, such as \code{\link[httr]{GET}}, \code{\link[httr]{POST}}, or \code{\link[httr]{DELETE}}.}
}
\value{
A list of class \dQuote{dataverse_group}.
}
\description{
\dots
}
\details{
\dots
}
\examples{
\dontrun{
# create a group
g <- create_group("my_dataverse", "testgroup", "aGroupName", "this is a test group")

# update group details
update_group(g, description = "this needs a new description")

# get group details
get_group(g)

# list all groups
list_groups("my_dataverse")

# delete group
delete_group(g)
}
}
\seealso{
\code{\link{list_groups}}, \code{\link{create_group}}, \code{\link{update_group}}, \code{\link{add_roles_to_group}}, \code{\link{delete_group}}
}

