% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/native_roles.R
\name{get_role}
\alias{get_role}
\title{Get role}
\usage{
get_role(role, key = Sys.getenv("DATAVERSE_KEY"),
  server = Sys.getenv("DATAVERSE_SERVER"), ...)
}
\arguments{
\item{role}{A character string specifying a role, or an object of class \dQuote{dataverse_role}.}

\item{key}{A character string specifying a Dataverse server API key. If one is not specified, functions calling authenticated API endpoints will fail. Keys can be specified atomically or globally using \code{Sys.setenv("DATAVERSE_KEY" = "examplekey")}.}

\item{server}{A character string specifying a Dataverse server. There are multiple Dataverse installations, but the defaults is to use the Harvard Dataverse. This can be modified atomically or globally using \code{Sys.setenv("DATAVERSE_SERVER" = "dataverse.example.com")}.}

\item{...}{Additional arguments passed to an HTTP request function, such as \code{\link[httr]{GET}}, \code{\link[httr]{POST}}, or \code{\link[httr]{DELETE}}.}
}
\value{
A list.
}
\description{
Retrieve a Dataverse role
}
\details{
Once created using \code{\link{create_role}}, this function can be used to retrieve roles for a Dataverse. \code{\link{delete_role}} can delete a role.
}
\examples{
\dontrun{}
}
\seealso{
\code{\link{create_role}}, \code{\link{delete_role}}
}

