% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SWORD.R
\name{initiate_dataset}
\alias{initiate_dataset}
\title{Allowed fields are:
\dQuote{abstract}, \dQuote{accessRights}, \dQuote{accrualMethod},
\dQuote{accrualPeriodicity}, \dQuote{accrualPolicy}, \dQuote{alternative},
\dQuote{audience}, \dQuote{available}, \dQuote{bibliographicCitation},
\dQuote{conformsTo}, \dQuote{contributor}, \dQuote{coverage}, \dQuote{created},
\dQuote{creator}, \dQuote{date}, \dQuote{dateAccepted}, \dQuote{dateCopyrighted},
\dQuote{dateSubmitted}, \dQuote{description}, \dQuote{educationLevel}, \dQuote{extent},
\dQuote{format}, \dQuote{hasFormat}, \dQuote{hasPart}, \dQuote{hasVersion},
\dQuote{identifier}, \dQuote{instructionalMethod}, \dQuote{isFormatOf},
\dQuote{isPartOf}, \dQuote{isReferencedBy}, \dQuote{isReplacedBy}, \dQuote{isRequiredBy},
\dQuote{issued}, \dQuote{isVersionOf}, \dQuote{language}, \dQuote{license},
\dQuote{mediator}, \dQuote{medium}, \dQuote{modified}, \dQuote{provenance},
\dQuote{publisher}, \dQuote{references}, \dQuote{relation}, \dQuote{replaces},
\dQuote{requires}, \dQuote{rights}, \dQuote{rightsHolder}, \dQuote{source},
\dQuote{spatial}, \dQuote{subject}, \dQuote{tableOfContents}, \dQuote{temporal},
\dQuote{title}, \dQuote{type}, and \dQuote{valid}.}
\usage{
initiate_dataset(dataverse, body, key = Sys.getenv("DATAVERSE_KEY"),
  server = Sys.getenv("DATAVERSE_SERVER"), ...)
}
\description{
Allowed fields are:
\dQuote{abstract}, \dQuote{accessRights}, \dQuote{accrualMethod},
\dQuote{accrualPeriodicity}, \dQuote{accrualPolicy}, \dQuote{alternative},
\dQuote{audience}, \dQuote{available}, \dQuote{bibliographicCitation},
\dQuote{conformsTo}, \dQuote{contributor}, \dQuote{coverage}, \dQuote{created},
\dQuote{creator}, \dQuote{date}, \dQuote{dateAccepted}, \dQuote{dateCopyrighted},
\dQuote{dateSubmitted}, \dQuote{description}, \dQuote{educationLevel}, \dQuote{extent},
\dQuote{format}, \dQuote{hasFormat}, \dQuote{hasPart}, \dQuote{hasVersion},
\dQuote{identifier}, \dQuote{instructionalMethod}, \dQuote{isFormatOf},
\dQuote{isPartOf}, \dQuote{isReferencedBy}, \dQuote{isReplacedBy}, \dQuote{isRequiredBy},
\dQuote{issued}, \dQuote{isVersionOf}, \dQuote{language}, \dQuote{license},
\dQuote{mediator}, \dQuote{medium}, \dQuote{modified}, \dQuote{provenance},
\dQuote{publisher}, \dQuote{references}, \dQuote{relation}, \dQuote{replaces},
\dQuote{requires}, \dQuote{rights}, \dQuote{rightsHolder}, \dQuote{source},
\dQuote{spatial}, \dQuote{subject}, \dQuote{tableOfContents}, \dQuote{temporal},
\dQuote{title}, \dQuote{type}, and \dQuote{valid}.
}

