% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native_roles_assignments.R
\name{get_assignments}
\alias{get_assignments}
\alias{assign_role}
\alias{delete_assignment}
\title{Role assignments}
\usage{
get_assignments(dataverse, key = Sys.getenv("DATAVERSE_KEY"),
  server = Sys.getenv("DATAVERSE_SERVER"), ...)

assign_role(dataverse, assignee, role, key = Sys.getenv("DATAVERSE_KEY"),
  server = Sys.getenv("DATAVERSE_SERVER"), ...)

delete_assignment(dataverse, assignment, key = Sys.getenv("DATAVERSE_KEY"),
  server = Sys.getenv("DATAVERSE_SERVER"), ...)
}
\arguments{
\item{dataverse}{A character string specifying a Dataverse name or an object of class \dQuote{dataverse}.}

\item{key}{A character string specifying a Dataverse server API key. If one is not specified, functions calling authenticated API endpoints will fail. Keys can be specified atomically or globally using \code{Sys.setenv("DATAVERSE_KEY" = "examplekey")}.}

\item{server}{A character string specifying a Dataverse server. There are multiple Dataverse installations, but the defaults is to use the Harvard Dataverse. This can be modified atomically or globally using \code{Sys.setenv("DATAVERSE_SERVER" = "dataverse.example.com")}.}

\item{...}{Additional arguments passed to an HTTP request function, such as \code{\link[httr]{GET}}, \code{\link[httr]{POST}}, or \code{\link[httr]{DELETE}}.}

\item{assignee}{\dots}

\item{role}{\dots}

\item{assignment}{\dots}
}
\value{
A list of objects of class \dQuote{dataverse_role_assignment}.
}
\description{
Get, set, and delete role assignments for a Dataverse
}
\details{
\code{assign_role} assigns a Dataverse role to a user.
}
\examples{
\dontrun{
# get role assignments
get_assignments("my_dataverse")
}
}
\seealso{
\code{\link{create_role}}
}
