% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native_role_groups.R
\name{create_group}
\alias{create_group}
\alias{update_group}
\alias{list_groups}
\alias{get_group}
\alias{delete_group}
\title{Role groups}
\usage{
create_group(dataverse, alias, name, description,
  key = Sys.getenv("DATAVERSE_KEY"),
  server = Sys.getenv("DATAVERSE_SERVER"), ...)

update_group(group, name, description, dataverse,
  key = Sys.getenv("DATAVERSE_KEY"),
  server = Sys.getenv("DATAVERSE_SERVER"), ...)

list_groups(dataverse, key = Sys.getenv("DATAVERSE_KEY"),
  server = Sys.getenv("DATAVERSE_SERVER"), ...)

get_group(group, dataverse, key = Sys.getenv("DATAVERSE_KEY"),
  server = Sys.getenv("DATAVERSE_SERVER"), ...)

delete_group(group, dataverse, key = Sys.getenv("DATAVERSE_KEY"),
  server = Sys.getenv("DATAVERSE_SERVER"), ...)
}
\arguments{
\item{dataverse}{A character string specifying a Dataverse name or an object of class \dQuote{dataverse}.}

\item{alias}{A character string specifying an alias for the role group.}

\item{name}{A character string specifying the name of the role group.}

\item{description}{A character string specifying a description of the role group.}

\item{key}{A character string specifying a Dataverse server API key. If one is not specified, functions calling authenticated API endpoints will fail. Keys can be specified atomically or globally using \code{Sys.setenv("DATAVERSE_KEY" = "examplekey")}.}

\item{server}{A character string specifying a Dataverse server. There are multiple Dataverse installations, but the defaults is to use the Harvard Dataverse. This can be modified atomically or globally using \code{Sys.setenv("DATAVERSE_SERVER" = "dataverse.example.com")}.}

\item{...}{Additional arguments passed to an HTTP request function, such as \code{\link[httr]{GET}}, \code{\link[httr]{POST}}, or \code{\link[httr]{DELETE}}.}

\item{group}{An object of class \dQuote{dataverse_group}, or a character string containing the \dQuote{alias} for a group. If a character string, \code{dataverse} must also be specified as the alias is local to a given Dataverse.}
}
\value{
A list of class \dQuote{dataverse_group}.
}
\description{
Create, update, and delete role groups
}
\details{
Role groups can be used to organize roles for a Dataverse.
}
\examples{
\dontrun{
# create a group
g <- create_group("my_dataverse", "testgroup", "aGroupName", "this is a test group")

# update group details
update_group(g, description = "this needs a new description")

# get group details
get_group(g)

# list all groups
list_groups("my_dataverse")

# delete group
delete_group(g)
}
}
\seealso{
\code{\link{add_roles_to_group}}
}
