% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/native_dataset.R
\name{dataset_versions}
\alias{dataset_versions}
\title{Dataset versions}
\usage{
dataset_versions(dataset, key = Sys.getenv("DATAVERSE_KEY"),
  server = Sys.getenv("DATAVERSE_SERVER"), ...)
}
\arguments{
\item{dataset}{An integer specifying a dataset identification number or an object of class \dQuote{dataverse_dataset}. The identification number is the dataset's persistent identification number (not the integer specifying a specific version of the dataset, such as returned by \code{\link{dataset_versions}}).}

\item{key}{A character string specifying a Dataverse server API key. If one is not specified, functions calling authenticated API endpoints will fail. Keys can be specified atomically or globally using \code{Sys.setenv("DATAVERSE_KEY" = "examplekey")}.}

\item{server}{A character string specifying a Dataverse server. There are multiple Dataverse installations, but the defaults is to use the Harvard Dataverse. This can be modified atomically or globally using \code{Sys.setenv("DATAVERSE_SERVER" = "dataverse.example.com")}.}
}
\value{
A list of class \dQuote{dataverse_dataset_version}.
}
\description{
View versions of a dataset
}
\details{
This returns a list of objects of all versions of a dataset, including metadata. This can be used as a first step for retrieving older versions of files or datasets.
}
\examples{
\dontrun{}
}
\seealso{
\code{\link{get_dataset}}, \code{\link{dataset_files}}, \code{\link{publish_dataset}}
}

