% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R, R/get_file_by_id.R
\name{get_file}
\alias{get_file}
\alias{get_file_by_id}
\title{Download File(s)}
\usage{
get_file(
  file,
  dataset = NULL,
  format = c("original", "bundle"),
  vars = NULL,
  key = Sys.getenv("DATAVERSE_KEY"),
  server = Sys.getenv("DATAVERSE_SERVER"),
  ...
)

get_file_by_id(
  fileid,
  dataset = NULL,
  server,
  format = c("original", "RData", "prep", "bundle"),
  vars,
  archival = NULL,
  key = Sys.getenv("DATAVERSE_KEY"),
  ...
)
}
\arguments{
\item{file}{An integer specifying a file identifier; or a vector of integers
specifying file identifiers; or, if \code{doi} is specified, a character string
specifying a file name within the DOI-identified dataset; or an object of
 class \dQuote{dataverse_file} as returned by \code{\link{dataset_files}}.}

\item{format}{A character string specifying a file format for download.
by default, this is \dQuote{original} (the original file format). If `NULL`,
no query is added, so ingested files are returned in their ingested TSV form.
For tabular datasets, the option \dQuote{bundle} downloads the bundle
of the original and archival versions, as well as the documentation.
See <https://guides.dataverse.org/en/latest/api/dataaccess.html> for details.}

\item{vars}{A character vector specifying one or more variable names, used to
extract a subset of the data.}

\item{key}{A character string specifying a Dataverse server API key. If one
is not specified, functions calling authenticated API endpoints will fail.
Keys can be specified atomically or globally using
\code{Sys.setenv("DATAVERSE_KEY" = "examplekey")}.}

\item{server}{A character string specifying a Dataverse server. There are
multiple Dataverse installations, but the defaults is to use the Harvard
Dataverse (`server = "dataverse.harvard.edu"`). This can be modified atomically
or globally using \code{Sys.setenv("DATAVERSE_SERVER" = "dataverse.example.com")}.}

\item{...}{Additional arguments passed to an HTTP request function, such as
\code{\link[httr]{GET}}, \code{\link[httr]{POST}}, or
\code{\link[httr]{DELETE}}.}

\item{fileid}{A numeric ID internally used for `get_file_by_id`}

\item{archival}{If a ingested (.tab) version is available, download
the ingested archival version or not?}
}
\value{
\code{get_file} returns a raw vector (or list of raw vectors,
  if \code{length(file) > 1}), which can be saved locally with the `writeBin`
  function.  To load datasets into the R environment dataframe, see
  \link{get_dataframe_by_name}.
}
\description{
Download Dataverse File(s). `get_file` internally calls
 `get_file_by_id`.
}
\details{
This function provides access to data files from a Dataverse entry.
}
\examples{
\dontrun{

# 1. Two-steps: Find ID from get_dataset
d1 <- get_dataset("doi:10.7910/DVN/ARKOTI", server = "dataverse.harvard.edu")
f1 <- get_file(d1$files$id[1], server = "dataverse.harvard.edu")

# 2. Using filename and dataverse
f2 <- get_file("constructionData.tab",
               "doi:10.7910/DVN/ARKOTI",
               server = "dataverse.harvard.edu")

# 3. Based on "dataverse_file" object
f3_dvf <- dataset_files(2692151, server = "dataverse.harvard.edu")
f3 <- get_file(f3_dvf[[2]], server = "dataverse.harvard.edu")

# 4. Retrieve multiple raw data in list
f4_vec <- get_dataset("doi:10.7910/DVN/CXOB4K",
                        server = "dataverse.harvard.edu")$files$id
f4 <- get_file(f4_vec,
               server = "dataverse.harvard.edu")
length(f4)

# Write binary files.
# The appropriate file extension needs to be assigned by the user.
writeBin(f2, "constructionData.tab")
writeBin(f4, "dataverse_download.zip")
writeBin(f4[[1]], "Appendices.docx")

}

}
\seealso{
To load the objects as datasets \link{get_dataframe_by_name}.
}
