% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_as_dataframe.R
\name{get_dataframe_by_name}
\alias{get_dataframe_by_name}
\alias{get_dataframe_by_id}
\alias{get_dataframe_by_doi}
\title{Get file from dataverse and convert it into a dataframe or tibble}
\usage{
get_dataframe_by_name(file, dataset = NULL, FUN = NULL, original = FALSE, ...)

get_dataframe_by_id(fileid, FUN = NULL, original = FALSE, ...)

get_dataframe_by_doi(filedoi, FUN = NULL, original = FALSE, ...)
}
\arguments{
\item{file}{to be passed on to get_file}

\item{dataset}{to be passed on to get_file}

\item{FUN}{The function to used for reading in the raw dataset. This user
must choose the appropriate function: for example if the target is a .rds
file, then \code{FUN} should be \code{readRDS} or \code{readr::read_rds}.}

\item{original}{A logical, defaulting to TRUE. Whether to read the ingested,
archival version of the dataset if one exists. The archival versions are tab-delimited
\code{.tab} files so if \code{original = FALSE}, \code{FUN} is set to \code{readr::read_tsv}.
If functions to read the original version is available, then \code{original = TRUE}
with a specified \code{FUN} is better.}

\item{...}{
  Arguments passed on to \code{\link[=get_file]{get_file}}
  \describe{
    \item{\code{format}}{A character string specifying a file format for download.
by default, this is \dQuote{original} (the original file format). If \code{NULL},
no query is added, so ingested files are returned in their ingested TSV form.
For tabular datasets, the option \dQuote{bundle} downloads the bundle
of the original and archival versions, as well as the documentation.
See \url{https://guides.dataverse.org/en/latest/api/dataaccess.html} for details.}
    \item{\code{vars}}{A character vector specifying one or more variable names, used to
extract a subset of the data.}
    \item{\code{key}}{A character string specifying a Dataverse server API key. If one
is not specified, functions calling authenticated API endpoints will fail.
Keys can be specified atomically or globally using
\code{Sys.setenv("DATAVERSE_KEY" = "examplekey")}.}
    \item{\code{server}}{A character string specifying a Dataverse server. There are
multiple Dataverse installations, but the defaults is to use the Harvard
Dataverse (\code{server = "dataverse.harvard.edu"}). This can be modified atomically
or globally using \code{Sys.setenv("DATAVERSE_SERVER" = "dataverse.example.com")}.}
  }}

\item{fileid}{A numeric ID internally used for \code{get_file_by_id}}

\item{filedoi}{A DOI for a single file (not the entire dataset), of the form
\code{"10.70122/FK2/PPIAXE/MHDB0O"} or \code{"doi:10.70122/FK2/PPIAXE/MHDB0O"}}
}
\description{
\code{get_dataframe_by_id}, if you know the numeric ID of the dataset, or instead
\code{get_dataframe_by_name} if you know the filename and doi. The dataset
}
\examples{
library(readr)

# load dataset from file name and dataverse DOI
csv_tab <- get_dataframe_by_name(
  file = "roster-bulls-1996.tab",
  dataset = "doi:10.70122/FK2/HXJVJU",
  server = "demo.dataverse.org")

# or a Stata dta
stata_df <- get_dataframe_by_name(
  file = "nlsw88.tab",
  dataset = "doi:10.70122/FK2/PPIAXE",
  server = "demo.dataverse.org")

# To use the original version, or for non-ingested data,
# please specify `orginal = TRUE` and specify a function in FUN

stata_df <- get_dataframe_by_name(
  file = "nlsw88.tab",
  dataset = "doi:10.70122/FK2/PPIAXE",
  server = "demo.dataverse.org",
  original = TRUE,
  FUN = haven::read_dta)

rds_df <- get_dataframe_by_name(
  file = "nlsw88_rds-export.rds",
  dataset = "doi:10.70122/FK2/PPIAXE",
  server = "demo.dataverse.org",
  FUN = readr::read_rds)

# equivalently, if you know the DOI
stata_df <- get_dataframe_by_doi(
  filedoi = "10.70122/FK2/PPIAXE/MHDB0O",
  server = "demo.dataverse.org",
  original = TRUE,
  FUN = haven::read_dta
)
}
