% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_ingested}
\alias{is_ingested}
\title{Identify if file is an ingested file}
\usage{
is_ingested(fileid, server = Sys.getenv("DATAVERSE_SERVER"))
}
\arguments{
\item{fileid}{A numeric fileid or file-specific DOI}
}
\description{
Identify if file is an ingested file
}
\examples{
# https://demo.dataverse.org/file.xhtml?persistentId=doi:10.70122/FK2/X5MUPQ/T0KKUZ
# nlsw88.tab
is_ingested(fileid = "doi:10.70122/FK2/X5MUPQ/T0KKUZ",
            server = "demo.dataverse.org")

# nlsw88_rds-export.rds
is_ingested(fileid = "doi:10.70122/FK2/PPIAXE/SUCFNI",
            server = "demo.dataverse.org")

}
